Change CPU struct design
Since the registers are going to be accessed and modified a lot, it
makes more sense for them to be stored as direct attributes instead of
being stored in an array. The registers being enums does not make much
sense because modifying and matching the values would generate a lot of
boilerplate code.